Enable more slots in DL (in theory: both TDD periods)
We cannot really enable slots in both TDD periods, because the way we
receive HARQ feedback is not ready and OAI-UE does not seem to support
it.

Note: the slots to "enable" depend on the TDD configuration, so instead
of configuring it statically, we should read it from the configuration
file.